+++ app/_layout.js
@@ -1,0 +1,12 @@
+import { Stack } from "expo-router";
+
+export default function Layout() {
+  return (
+    <>
+      <Stack>
+        <Stack.Screen name="navBar" />
+        <Stack.Screen name="index" />
+      </Stack>
+    </>
+  );
+}
