+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "notifications",
   "version": "1.0.0",
-  "main": "expo/AppEntry.js",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
